feat: created the ReportFrom and CreateForm components
