+++ components/ReportTable.js
@@ -1,0 +1,5 @@
+export default function ReportTable() {
+    return (
+        
+    )
+}
